// src/types/registerRequestDTO.ts
/**
 * Generated by orval v7.16.1 🍺
 * Do not edit manually.
 * SE114.P12
 * API cho dự án nhập môn ứng dụng di dộng
 * OpenAPI spec version: 1.0
 */

export interface RegisterRequestDTO {
  fullname: string;
  /** @pattern ^(\+84|0)(3[2-9]|5[2-9]|7[0|6-9]|8[1-9]|9[0-4|6-9])[0-9]{7}$ */
  phone: string;
  email?: string;
  username: string;
  password: string;
}